Brave new world

+++ src/client/cl_media.h
@@ -24,7 +24,7 @@
 #include "cl_types.h"
 
 GList *Cl_Mapshots(const char *mapname);
-void Cl_LoadingProgress(uint16_t percent, const char *file);
+void Cl_LoadingProgress(int32_t percent, const char *file);
 
 #ifdef __CL_LOCAL_H__
