+++ src/client/cl_predict.h
@@ -25,7 +25,7 @@
 
 int32_t Cl_PointContents(const vec3_t point);
 cm_trace_t Cl_Trace(const vec3_t start, const vec3_t end, const vec3_t mins, const vec3_t maxs,
-                    const uint16_t skip, const int32_t contents);
+                    const int32_t skip, const int32_t contents);
 
 #ifdef __CL_LOCAL_H__
 void Cl_PredictMovement(void);
